third_party/WebKit/LayoutTests/editing/inserting/insert_between_block_quote.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<blockquote>foo</blockquote>',
            '|<br>',
            '<blockquote>bar</blockquote>',
        '</div>',
    ].join(''),
    'insertText baz',
    [
        '<div contenteditable>',
            '<blockquote>foo</blockquote>',
            'baz|<br>',
            '<blockquote>bar</blockquote>',
        '</div>',
    ].join('')),
    'insert text between BLOCKQUOTE');
</script>
